send bitcoin part 2

+++ src/main/java/com/byow/wallet/byow/utils/BitcoinFormatter.java
@@ -1,11 +1,12 @@
 package com.byow.wallet.byow.utils;
 
+import java.math.BigDecimal;
 import java.text.DecimalFormat;
 import java.text.DecimalFormatSymbols;
 import java.util.Locale;
 
 public class BitcoinFormatter {
-    public static String format(double number) {
+    public static String format(BigDecimal number) {
         DecimalFormatSymbols symbols = new DecimalFormatSymbols(Locale.ROOT);
         symbols.setDecimalSeparator('.');
         symbols.setGroupingSeparator(',');
@@ -13,6 +14,6 @@ public class BitcoinFormatter {
         formatter.setGroupingUsed(false);
         formatter.setMinimumFractionDigits(8);
         formatter.setMaximumFractionDigits(8);
-        return formatter.format(number);
+        return formatter.format(number.doubleValue());
     }
 }
